typescriptに関するライブラリを入れて、コンパイルした。
