+++ formcoder-api/src/index.js
@@ -1,11 +1,9 @@
-const express = require("express");
-const app = express();
-const port = 3000;
-
-app.get("/", (req, res) => {
-  res.send("Hello world!");
+var express = require("express");
+var app = express();
+var port = 3000;
+app.get("/", function (req, res) {
+    res.send("Hello world!");
 });
-
-app.listen(port, () => {
-  console.log(`Example app listening on port ${port}`);
+app.listen(port, function () {
+    console.log("Example app listening on port ".concat(port));
 });
