+++ formcoder-api/src/index.ts
@@ -1,0 +1,11 @@
+const express = require("express");
+const app = express();
+const port = 3000;
+
+app.get("/", (req, res) => {
+  res.send("Hello world!");
+});
+
+app.listen(port, () => {
+  console.log(`Example app listening on port ${port}`);
+});
